requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mongomock==4.3.0
mongomock-motor==0.0.35
motor==3.7.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
pytz==2025.2
sentinels==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.1
trio==0.29.0
typing_extensions==4.13.0
uvicorn==0.34.0
